refactor: Rename `mod cache` to `mod entity_cache`
